# openapi.json
{"openapi":"3.1.0","info":{"title":"Nifty 100 Financial Intelligence API","description":"REST API for the Nifty 100 Financial Intelligence Platform.","version":"1.0.0"},"paths":{"/api/companies":{"get":{"tags":["Companies"],"summary":"List Companies","description":"Return the company universe with sector information.","operationId":"list_companies_api_companies_get","parameters":[{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/companies/{company_id}":{"get":{"tags":["Companies"],"summary":"Company Detail","description":"Return company master information and sector classification.","operationId":"company_detail_api_companies__company_id__get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/companies/{company_id}/ratios":{"get":{"tags":["Companies"],"summary":"Company Ratios","description":"Return financial ratios for a company and period.","operationId":"company_ratios_api_companies__company_id__ratios_get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}},{"name":"year","in":"query","required":false,"schema":{"type":"string","default":"2024-03","title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/companies/{company_id}/valuation":{"get":{"tags":["Companies"],"summary":"Company Valuation","description":"Return market-cap and valuation metrics.","operationId":"company_valuation_api_companies__company_id__valuation_get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}},{"name":"year","in":"query","required":false,"schema":{"type":"string","default":"2024-03","title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/companies/{company_id}/prices":{"get":{"tags":["Companies"],"summary":"Company Prices","description":"Return most recent available stock-price observations.","operationId":"company_prices_api_companies__company_id__prices_get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":30,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/companies/{company_id}/signals":{"get":{"tags":["Companies"],"summary":"Company Signals","description":"Return generated analytical pros/cons signals.","operationId":"company_signals_api_companies__company_id__signals_get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/sectors":{"get":{"tags":["Analytics"],"summary":"Sector Summary","description":"Return company count and index weight by broad sector.","operationId":"sector_summary_api_analytics_sectors_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/analytics/sectors/{sector_name}":{"get":{"tags":["Analytics"],"summary":"Sector Companies","description":"Return companies belonging to a broad sector.","operationId":"sector_companies_api_analytics_sectors__sector_name__get","parameters":[{"name":"sector_name","in":"path","required":true,"schema":{"type":"string","title":"Sector Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/latest-ratios":{"get":{"tags":["Analytics"],"summary":"Latest Ratios","description":"Return latest meaningful annual ratio row per company.","operationId":"latest_ratios_api_analytics_latest_ratios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/analytics/valuation":{"get":{"tags":["Analytics"],"summary":"Valuation Summary","description":"Return latest available valuation data per company.","operationId":"valuation_summary_api_analytics_valuation_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/analytics/signals":{"get":{"tags":["Analytics"],"summary":"Signal Summary","description":"Return generated pro/con signal counts by company.","operationId":"signal_summary_api_analytics_signals_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/analytics/clusters":{"get":{"tags":["Analytics"],"summary":"Cluster Analytics","description":"Return Day 36-37 company cluster labels and cluster profiles.","operationId":"cluster_analytics_api_analytics_clusters_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/analytics/outliers":{"get":{"tags":["Analytics"],"summary":"Outlier Analytics","description":"Return Day 37 statistical outlier report.","operationId":"outlier_analytics_api_analytics_outliers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/analytics/portfolio":{"get":{"tags":["Analytics"],"summary":"Portfolio Statistics","description":"Return Day 37 portfolio-level KPI statistics.","operationId":"portfolio_statistics_api_analytics_portfolio_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["System"],"summary":"Root","description":"Return API identification information.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["System"],"summary":"Health","description":"Return API and database health.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}